+++ src/components/inputs/InputComponents/index.ts
@@ -1,0 +1,11 @@
+import DateInput from '@/components/inputs/InputComponents/DateInput';
+import TagInput from '@/components/inputs/InputComponents/TagInput';
+import TextInput from '@/components/inputs/InputComponents/TextInput';
+import Textarea from '@/components/inputs/InputComponents/Textarea';
+
+export const InputComponents = Object.assign({
+  TextInput: TextInput,
+  DateInput: DateInput,
+  TagInput: TagInput,
+  Textarea: Textarea,
+});
