+++ src/components/modal/Confirm.tsx
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import { Button } from '@/components/buttons';
+import { ConfirmComponents } from '@/components/modal/ConfirmComponents';
 
 interface Props {
   onCloseModal: () => void;
@@ -18,34 +18,7 @@ function Confirm({ onCloseModal, children }: Props) {
   );
 }
 
-function CheckConfirm({ onCloseModal }: Props) {
-  return (
-    <>
-      <span className='relative bottom-10'>비밀번호가 일치하지 않습니다.</span>
-      <div className='absolute bottom-0 tablet:right-0'>
-        <Button size='lg' onClick={onCloseModal}>
-          확인
-        </Button>
-      </div>
-    </>
-  );
-}
-
-function DeleteConfirm({ onCloseModal }: Props) {
-  return (
-    <>
-      <span className='relative bottom-10'>컬럼의 모든 카드가 삭제됩니다.</span>
-      <div className='absolute bottom-0 flex gap-10 tablet:right-0'>
-        <Button.Secondary size='lg' onClick={onCloseModal}>
-          취소
-        </Button.Secondary>
-        <Button size='lg'>삭제</Button>
-      </div>
-    </>
-  );
-}
-
-Confirm.CheckConfirm = CheckConfirm;
-Confirm.DeleteConfirm = DeleteConfirm;
+Confirm.CheckConfirm = ConfirmComponents.CheckConfirm;
+Confirm.DeleteConfirm = ConfirmComponents.DeleteConfirm;
 
 export default Confirm;
