+++ src/components/modal/ConfirmComponents/index.ts
@@ -1,0 +1,7 @@
+import CheckConfirm from '@/components/modal/ConfirmComponents/CheckConfirm';
+import DeleteConfirm from '@/components/modal/ConfirmComponents/DeleteConfirm';
+
+export const ConfirmComponents = Object.assign({
+  CheckConfirm: CheckConfirm,
+  DeleteConfirm: DeleteConfirm,
+});
