+++ src/components/modal/FormComponents/index.ts
@@ -1,0 +1,9 @@
+import ColumnForm from '@/components/modal/FormComponents/ColumnForm';
+import DashboardForm from '@/components/modal/FormComponents/DashboardForm';
+import TodoForm from '@/components/modal/FormComponents/TodoForm';
+
+export const FormComponents = Object.assign({
+  DashboardForm: DashboardForm,
+  TodoForm: TodoForm,
+  ColumnForm: ColumnForm,
+});
